[TSan] Move EscapeAnalysis to lib/Analysis and enable by default
Address reviewer comments and prepare the pass for upstreaming:
* Move EscapeAnalysis from Transforms/Instrumentation to llvm/Analysis, as it is a pure analysis pass.
* Enable the optimization by default (-tsan-use-escape-analysis=true).
* Add detailed documentation to the flag explaining safety guarantees.
* Clean up custom lit test configurations as they are no longer needed.
* Include a small drive-by fix for `sanitizer-ld.c` to prevent test failures when the build directory path contains the substring "tsan".
