+++ clang/test/Driver/sanitizer-ld.c
@@ -1386,7 +1386,8 @@
 // RUN:     --sysroot=%S/Inputs/basic_linux_tree \
 // RUN:   | %{filecheck} --check-prefix=CHECK-RELOCATABLE-LINK-ASAN-UBSAN-RTLIB
 //
-// CHECK-RELOCATABLE-LINK-ASAN-UBSAN-RTLIB-NOT: "{{.*}}(asan|ubsan){{.*}}"
+// CHECK-RELOCATABLE-LINK-ASAN-UBSAN-RTLIB-NOT: libclang_rt.asan
+// CHECK-RELOCATABLE-LINK-ASAN-UBSAN-RTLIB-NOT: libclang_rt.ubsan
 
 // RUN: %clang -fsanitize=address -r -fsanitize-link-runtime -### %s 2>&1 \
 // RUN:     --target=x86_64-unknown-linux -fuse-ld=ld \
@@ -1394,7 +1395,7 @@
 // RUN:     --sysroot=%S/Inputs/basic_linux_tree \
 // RUN:   | FileCheck %s --check-prefix=CHECK-RELOCATABLE-LINK-FORCE-LINK-ASAN
 //
-// CHECK-RELOCATABLE-LINK-FORCE-LINK-ASAN: "{{.*}}asan{{.*}}"
+// CHECK-RELOCATABLE-LINK-FORCE-LINK-ASAN: libclang_rt.asan
 
 // RUN: %clang -fsanitize=thread -r -### %s 2>&1 \
 // RUN:     --target=x86_64-unknown-linux -fuse-ld=ld \
@@ -1402,7 +1403,7 @@
 // RUN:     --sysroot=%S/Inputs/basic_linux_tree \
 // RUN:   | %{filecheck} --check-prefix=CHECK-RELOCATABLE-LINK-TSAN-RTLIB
 //
-// CHECK-RELOCATABLE-LINK-TSAN-RTLIB-NOT: "{{.*}}tsan{{.*}}"
+// CHECK-RELOCATABLE-LINK-TSAN-RTLIB-NOT: libclang_rt.tsan
 
 // RUN: %clang -fsanitize=fuzzer,address -shared-libsan -### %s 2>&1 \
 // RUN:     --target=x86_64-unknown-linux -fuse-ld=ld \
