+++ compiler-rt/test/tsan/CMakeLists.txt
@@ -18,7 +18,6 @@ endif()
 set(TSAN_DYNAMIC_TEST_DEPS ${TSAN_TEST_DEPS})
 set(TSAN_TESTSUITES)
 set(TSAN_DYNAMIC_TESTSUITES)
-set(TSAN_ENABLE_ESCAPE_ANALYSIS "False") # Disable escape analysis by default
 
 if (NOT DEFINED TSAN_TEST_DEFLAKE_THRESHOLD)
   set(TSAN_TEST_DEFLAKE_THRESHOLD "10")
@@ -29,79 +28,47 @@ if(APPLE)
   darwin_filter_host_archs(TSAN_SUPPORTED_ARCH TSAN_TEST_ARCH)
 endif()
 
-# Unified function for generating TSAN test suites by architectures.
-# Arguments:
-#   OUT_LIST_VAR    - name of output list (for example, TSAN_TESTSUITES or TSAN_EA_TESTSUITES)
-#   SUFFIX_KIND     - string added to config suffix after "-${arch}" (for example, "" or "-escape")
-#   CONFIG_KIND     - string added to config name after "Config" (for example, "" or "Escape")
-#   ENABLE_EA       - "True"/"False" enable escape analysis
-function(tsan_generate_arch_suites OUT_LIST_VAR SUFFIX_KIND CONFIG_KIND ENABLE_EA)
-  foreach(arch ${TSAN_TEST_ARCH})
-    set(TSAN_ENABLE_ESCAPE_ANALYSIS "${ENABLE_EA}")
+foreach(arch ${TSAN_TEST_ARCH})
+  set(TSAN_TEST_APPLE_PLATFORM "osx")
+  set(TSAN_TEST_MIN_DEPLOYMENT_TARGET_FLAG "${DARWIN_osx_MIN_VER_FLAG}")
+  set(TSAN_TEST_APPLE_TARGET_IS_HOST ON)
+  pythonize_bool(TSAN_TEST_APPLE_TARGET_IS_HOST)
 
-    set(TSAN_TEST_APPLE_PLATFORM "osx")
-    set(TSAN_TEST_MIN_DEPLOYMENT_TARGET_FLAG "${DARWIN_osx_MIN_VER_FLAG}")
-    set(TSAN_TEST_APPLE_TARGET_IS_HOST ON)
-    pythonize_bool(TSAN_TEST_APPLE_TARGET_IS_HOST)
+  set(TSAN_TEST_TARGET_ARCH ${arch})
+  string(TOLOWER "-${arch}" TSAN_TEST_CONFIG_SUFFIX)
+  get_test_cc_for_arch(${arch} TSAN_TEST_TARGET_CC TSAN_TEST_TARGET_CFLAGS)
 
-    set(TSAN_TEST_TARGET_ARCH ${arch})
-    string(TOLOWER "-${arch}${SUFFIX_KIND}" TSAN_TEST_CONFIG_SUFFIX)
-    get_test_cc_for_arch(${arch} TSAN_TEST_TARGET_CC TSAN_TEST_TARGET_CFLAGS)
+  string(REPLACE ";" " " LIBDISPATCH_CFLAGS_STRING " ${COMPILER_RT_TEST_LIBDISPATCH_CFLAGS}")
+  string(APPEND TSAN_TEST_TARGET_CFLAGS ${LIBDISPATCH_CFLAGS_STRING})
 
-    string(REPLACE ";" " " LIBDISPATCH_CFLAGS_STRING " ${COMPILER_RT_TEST_LIBDISPATCH_CFLAGS}")
-    string(APPEND TSAN_TEST_TARGET_CFLAGS ${LIBDISPATCH_CFLAGS_STRING})
-
-    if (COMPILER_RT_HAS_MSSE4_2_FLAG)
-      string(APPEND TSAN_TEST_TARGET_CFLAGS " -msse4.2 ")
-    endif()
-
-    string(TOUPPER ${arch} ARCH_UPPER_CASE)
-    set(CONFIG_NAME ${ARCH_UPPER_CASE}Config${CONFIG_KIND})
-
-    configure_lit_site_cfg(
-            ${CMAKE_CURRENT_SOURCE_DIR}/lit.site.cfg.py.in
-            ${CMAKE_CURRENT_BINARY_DIR}/${CONFIG_NAME}/lit.site.cfg.py
-            MAIN_CONFIG
-            ${CMAKE_CURRENT_SOURCE_DIR}/lit.cfg.py
-    )
-    list(APPEND ${OUT_LIST_VAR} ${CMAKE_CURRENT_BINARY_DIR}/${CONFIG_NAME})
-
-    if(COMPILER_RT_TSAN_HAS_STATIC_RUNTIME)
-      # Dynamic runtime for corresponding variant
-      if("${SUFFIX_KIND}" STREQUAL "")
-        string(TOLOWER "-${arch}-${OS_NAME}-dynamic" TSAN_TEST_CONFIG_SUFFIX)
-        set(CONFIG_NAME ${ARCH_UPPER_CASE}${OS_NAME}DynamicConfig${CONFIG_KIND})
-        list(APPEND TSAN_DYNAMIC_TESTSUITES ${CMAKE_CURRENT_BINARY_DIR}/${CONFIG_NAME})
-      else()
-        string(TOLOWER "-${arch}-${OS_NAME}-dynamic${SUFFIX_KIND}" TSAN_TEST_CONFIG_SUFFIX)
-        set(CONFIG_NAME ${ARCH_UPPER_CASE}${OS_NAME}DynamicConfig${CONFIG_KIND})
-        # Track dynamic escape-analysis suites separately for a dedicated target.
-        list(APPEND TSAN_EA_DYNAMIC_TESTSUITES ${CMAKE_CURRENT_BINARY_DIR}/${CONFIG_NAME})
-      endif()
-      configure_lit_site_cfg(
-              ${CMAKE_CURRENT_SOURCE_DIR}/lit.site.cfg.py.in
-              ${CMAKE_CURRENT_BINARY_DIR}/${CONFIG_NAME}/lit.site.cfg.py
-              MAIN_CONFIG
-              ${CMAKE_CURRENT_SOURCE_DIR}/lit.cfg.py
-      )
-      list(APPEND ${OUT_LIST_VAR} ${CMAKE_CURRENT_BINARY_DIR}/${CONFIG_NAME})
-    endif()
-  endforeach()
-
-  # Propagate the assembled list to the parent scope
-  set(${OUT_LIST_VAR} "${${OUT_LIST_VAR}}" PARENT_SCOPE)
-  if(DEFINED TSAN_EA_DYNAMIC_TESTSUITES)
-    set(TSAN_EA_DYNAMIC_TESTSUITES "${TSAN_EA_DYNAMIC_TESTSUITES}" PARENT_SCOPE)
+  if (COMPILER_RT_HAS_MSSE4_2_FLAG)
+    string(APPEND TSAN_TEST_TARGET_CFLAGS " -msse4.2 ")
   endif()
-endfunction()
 
-# Default configuration
-set(TSAN_TESTSUITES)
-tsan_generate_arch_suites(TSAN_TESTSUITES "" "" "False")
+  string(TOUPPER ${arch} ARCH_UPPER_CASE)
+  set(CONFIG_NAME ${ARCH_UPPER_CASE}Config)
 
-# Enable escape analysis (check-tsan-escape-analysis target)
-set(TSAN_EA_TESTSUITES)
-tsan_generate_arch_suites(TSAN_EA_TESTSUITES "-escape" "Escape" "True")
+  configure_lit_site_cfg(
+    ${CMAKE_CURRENT_SOURCE_DIR}/lit.site.cfg.py.in
+    ${CMAKE_CURRENT_BINARY_DIR}/${CONFIG_NAME}/lit.site.cfg.py
+    MAIN_CONFIG
+    ${CMAKE_CURRENT_SOURCE_DIR}/lit.cfg.py
+    )
+  list(APPEND TSAN_TESTSUITES ${CMAKE_CURRENT_BINARY_DIR}/${CONFIG_NAME})
+
+  if(COMPILER_RT_TSAN_HAS_STATIC_RUNTIME)
+    string(TOLOWER "-${arch}-${OS_NAME}-dynamic" TSAN_TEST_CONFIG_SUFFIX)
+    set(CONFIG_NAME ${ARCH_UPPER_CASE}${OS_NAME}DynamicConfig)
+    configure_lit_site_cfg(
+      ${CMAKE_CURRENT_SOURCE_DIR}/lit.site.cfg.py.in
+      ${CMAKE_CURRENT_BINARY_DIR}/${CONFIG_NAME}/lit.site.cfg.py
+      MAIN_CONFIG
+      ${CMAKE_CURRENT_SOURCE_DIR}/lit.cfg.py
+      )
+    list(APPEND TSAN_DYNAMIC_TESTSUITES
+      ${CMAKE_CURRENT_BINARY_DIR}/${CONFIG_NAME})
+  endif()
+endforeach()
 
 # iOS and iOS simulator test suites
 # These are not added into "check-all", in order to run these tests, use
@@ -161,10 +128,6 @@ list(APPEND TSAN_TESTSUITES ${CMAKE_CURRENT_BINARY_DIR}/Unit)
 if(COMPILER_RT_TSAN_HAS_STATIC_RUNTIME)
   list(APPEND TSAN_DYNAMIC_TESTSUITES ${CMAKE_CURRENT_BINARY_DIR}/Unit/dynamic)
 endif()
-list(APPEND TSAN_EA_TESTSUITES ${CMAKE_CURRENT_BINARY_DIR}/Unit)
-if(COMPILER_RT_TSAN_HAS_STATIC_RUNTIME)
-  list(APPEND TSAN_EA_DYNAMIC_TESTSUITES ${CMAKE_CURRENT_BINARY_DIR}/Unit/dynamic)
-endif()
 
 add_lit_testsuite(check-tsan "Running ThreadSanitizer tests"
   ${TSAN_TESTSUITES}
@@ -176,18 +139,4 @@ if(COMPILER_RT_TSAN_HAS_STATIC_RUNTIME)
                     ${TSAN_DYNAMIC_TESTSUITES}
                     EXCLUDE_FROM_CHECK_ALL
                     DEPENDS ${TSAN_DYNAMIC_TEST_DEPS})
-endif()
-
-add_lit_testsuite(check-tsan-escape-analysis "Running ThreadSanitizer tests (escape analysis)"
-        ${TSAN_EA_TESTSUITES}
-        DEPENDS ${TSAN_TEST_DEPS})
-set_target_properties(check-tsan-escape-analysis PROPERTIES FOLDER "Compiler-RT Tests")
-
-# New target: dynamic + escape analysis
-if(COMPILER_RT_TSAN_HAS_STATIC_RUNTIME)
-  add_lit_testsuite(check-tsan-escape-analysis-dynamic "Running ThreadSanitizer tests (dynamic, escape analysis)"
-                    ${TSAN_EA_DYNAMIC_TESTSUITES}
-                    EXCLUDE_FROM_CHECK_ALL
-                    DEPENDS ${TSAN_DYNAMIC_TEST_DEPS})
-  set_target_properties(check-tsan-escape-analysis-dynamic PROPERTIES FOLDER "Compiler-RT Tests")
 endif()
