+++ llvm/lib/Analysis/EscapeAnalysis.cpp
@@ -11,7 +11,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "llvm/Transforms/Instrumentation/EscapeAnalysis.h"
+#include "llvm/Analysis/EscapeAnalysis.h"
 #include "llvm/ADT/SmallString.h"
 #include "llvm/ADT/Statistic.h"
 #include "llvm/Analysis/AliasAnalysis.h"
